Split Rpc handlers into seperate package
