+++ packages/client-rpc-handlers/README.md
@@ -1,0 +1,1 @@
+# @polkadot/client-rpc-handlers
