+++ packages/client-rpc-handlers/src/system.ts
@@ -1,11 +1,12 @@
-// Copyright 2017-2018 @polkadot/client authors & contributors
+// Copyright 2017-2018 @polkadot/client-rpc-handlers authors & contributors
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
 
 import { ChainInterface } from '@polkadot/client-chains/types';
-import { Config, Endpoint } from '../types';
+import { Config } from '@polkadot/client/types';
+import { Endpoint } from './types';
 
-import * as clientId from '../clientId';
+import * as clientId from '@polkadot/client/clientId';
 
 const systemChain = ({ chain }: Config) =>
   async (): Promise<string> =>
